2022.02.02 - 2 (Success in Unfollowing and Unsaving)

+++ app/src/main/java/ir/mahdiparastesh/instatools/list/ListUnf.kt
@@ -3,6 +3,7 @@ package ir.mahdiparastesh.instatools.list
 import android.annotation.SuppressLint
 import android.content.Intent
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.android.volley.Request
 import com.bumptech.glide.Glide
@@ -54,6 +55,14 @@ class ListUnf(val c: Main, private val f: PageUnf) : RecyclerView.Adapter<ListUn
             if (c.m.unfollowers == null) return@setOnClickListener
             if (!c.m.unfollowers!![h.layoutPosition].isPrivate)
                 unfollow(c.m.unfollowers!![h.layoutPosition])
+            else AlertDialog.Builder(c).apply {
+                setTitle(R.string.unfollow)
+                setMessage(R.string.unfollowPV)
+                setNegativeButton(R.string.no, null)
+                setPositiveButton(R.string.yes) { _, _ ->
+                    unfollow(c.m.unfollowers!![h.layoutPosition])
+                }
+            }.create().show()
         }
         vis(h.b.sep, i < itemCount - 1)
     }
@@ -63,8 +72,11 @@ class ListUnf(val c: Main, private val f: PageUnf) : RecyclerView.Adapter<ListUn
     private fun unfollow(unf: Unfollower) {
         Api<Rest>(
             c, Api.Type.UNFOLLOW.url.format(unf.id.toString()), Rest::class,
-            method = Request.Method.POST
+            PageUnf.theHandler, method = Request.Method.POST
         ) {
+            if (it.status != "ok") {
+                PageUnf.theHandler?.obtainMessage(PageUnf.Action.COULD_NOT.ordinal)?.sendToTarget()
+                return@Api; }
             c.pDao.deleteUnfollower(unf)
             val index = c.m.unfollowers!!.indexOf(unf)
             c.m.unfollowers!!.remove(unf)
